put every logic related to help in help command
