+++ src/cli/commands/help.js
@@ -4,7 +4,7 @@ import * as commands from './index.js';
 import * as constants from '../../constants.js';
 import type {Reporter} from '../../reporters/index.js';
 import type Config from '../../config.js';
-import {sortAlpha, hyphenate} from '../../util/misc.js';
+import {sortAlpha, hyphenate, camelCase} from '../../util/misc.js';
 const chalk = require('chalk');
 
 export function run(
@@ -17,24 +17,47 @@ export function run(
   const getDocsInfo = (name) => 'Visit ' + chalk.bold(getDocsLink(name)) + ' for documentation about this command.';
 
   if (args.length) {
-    const helpCommand = hyphenate(args[0]);
-    if (commands[helpCommand]) {
-      commander.on('--help', () => console.log('  ' + getDocsInfo(helpCommand) + '\n'));
-    }
-  } else {
-    commander.on('--help', () => {
-      console.log('  Commands:\n');
-      for (const name of Object.keys(commands).sort(sortAlpha)) {
-        if (commands[name].useless) {
-          continue;
+    const commandName = camelCase(args.shift());
+
+    if (commandName) {
+      const command = commands[commandName];
+
+      if (command) {
+        if (typeof command.setFlags === 'function') {
+          command.setFlags(commander);
         }
 
-        console.log(`    - ${hyphenate(name)}`);
+        const examples: Array<string> = (command && command.examples) || [];
+        if (examples.length) {
+          commander.on('--help', () => {
+            console.log('  Examples:\n');
+            for (const example of examples) {
+              console.log(`    $ yarn ${example}`);
+            }
+            console.log();
+          });
+        }
+        commander.on('--help', () => console.log('  ' + getDocsInfo(commandName) + '\n'));
+
+        commander.help();
+        return Promise.resolve();
       }
-      console.log('\n  Run `' + chalk.bold('yarn help COMMAND') + '` for more information on specific commands.');
-      console.log('  Visit ' + chalk.bold(getDocsLink()) + ' to learn more about Yarn.\n');
-    });
+    }
   }
+
+  commander.on('--help', () => {
+    console.log('  Commands:\n');
+    for (const name of Object.keys(commands).sort(sortAlpha)) {
+      if (commands[name].useless) {
+        continue;
+      }
+
+      console.log(`    - ${hyphenate(name)}`);
+    }
+    console.log('\n  Run `' + chalk.bold('yarn help COMMAND') + '` for more information on specific commands.');
+    console.log('  Visit ' + chalk.bold(getDocsLink()) + ' to learn more about Yarn.\n');
+  });
+
   commander.help();
   return Promise.resolve();
 }
